import product

+++ src/main/java/com/piseth/java/school/phoneshop/controller/BrandController.java
@@ -36,18 +36,18 @@ public class BrandController {
 	}
 	
 	@GetMapping("{id}")
-	public ResponseEntity<?> getById(@PathVariable("id") int id) throws ApiException{
+	public ResponseEntity<?> getById(@PathVariable("id") Long id) throws ApiException{
 		return ResponseEntity.ok(brandService.getById(id));
 	}
 	
 	@PutMapping("{id}")
-	public ResponseEntity<?> update(@PathVariable("id") int id, @RequestBody BrandDTO brandDTO) throws ApiException {
+	public ResponseEntity<?> update(@PathVariable("id") Long id, @RequestBody BrandDTO brandDTO) throws ApiException {
 		Brand brand = BrandMapper.INSTANCE.toEntity(brandDTO);
 		return ResponseEntity.ok(brandService.update(id, brand));
 	}
 	
 	@DeleteMapping("{id}")
-	public ResponseEntity<?> deleteById(@PathVariable("id") int id) throws ApiException{
+	public ResponseEntity<?> deleteById(@PathVariable("id") Long id) throws ApiException{
 		brandService.delete(id);
 		return ResponseEntity.ok().build();
 	}
