+++ src/main/java/com/piseth/java/school/phoneshop/exception/ResourceNotFoundException.java
@@ -11,9 +11,9 @@ import lombok.Setter;
 public class ResourceNotFoundException extends ApiException{
 	
 	private String resourceName;
-	private Integer resourceId;
+	private Long resourceId;
 	
-	public ResourceNotFoundException(String resourceName, Integer resourceId) {
+	public ResourceNotFoundException(String resourceName, Long resourceId) {
 		super(HttpStatus.NOT_FOUND, String.format("%s not found for id=%d", resourceName, resourceId));
 	}
