+++ src/main/java/com/piseth/java/school/phoneshop/repository/BrandRepository.java
@@ -8,11 +8,11 @@ import org.springframework.stereotype.Repository;
 import com.piseth.java.school.phoneshop.model.Brand;
 
 @Repository
-public interface BrandRepository extends JpaRepository<Brand, Integer>{
+public interface BrandRepository extends JpaRepository<Brand, Long>{
 	
 	boolean existsByName(String name);
 	
-	List<Brand> findByIdIn(List<Integer> ids);
+	List<Brand> findByIdIn(List<Long> ids);
 	
 	//List<Brand> findByActive(boolean isActive);
 	List<Brand> findByActiveTrue();
