+++ src/main/java/com/piseth/java/school/phoneshop/repository/ModelRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
 import com.piseth.java.school.phoneshop.model.Model;
 
-public interface ModelRepository extends JpaRepository<Model, Integer> , JpaSpecificationExecutor<Model>{
+public interface ModelRepository extends JpaRepository<Model, Long> , JpaSpecificationExecutor<Model>{
 
 }
